Create separate entry points for each platform
Each platform gets its own entry point, so there is no need to dynamically detect which platform is running.
