+++ index.js.flow
@@ -13,13 +13,13 @@ export * from './src/types/types.js'
 
 declare export var fakeUser: Object
 declare export var fakeUser1: Object
-declare export function makeBrowserIo(): EdgeIo
-declare export function makeFakeIos(count: number): Array<EdgeIo>
-declare export function makeNodeIo(path: string): EdgeIo
-declare export function makeReactNativeIo(): Promise<EdgeIo>
-
 declare export function destroyAllContexts(): mixed
+declare export function makeFakeIos(count: number): Array<EdgeIo>
 
+/**
+ * Initializes the Edge core library,
+ * automatically selecting the appropriate platform.
+ */
 declare export function makeEdgeContext(
   opts: EdgeContextOptions
 ): Promise<EdgeContext>
@@ -27,3 +27,8 @@ declare export function makeEdgeContext(
 declare export function makeFakeContexts(
   ...opts: Array<EdgeFakeContextOptions>
 ): Promise<Array<EdgeContext>>
+
+// System-specific io exports:
+declare export function makeBrowserIo(): EdgeIo
+declare export function makeNodeIo(path: string): EdgeIo
+declare export function makeReactNativeIo(): Promise<EdgeIo>
