--- src/internal.js
@@ -1,9 +1,0 @@
-// @flow
-
-// We are exporting some internal goodies for the CLI,
-// which makes use of some undocumented core features.
-// In the future we hope to minimize / reduce this
-
-export { hmacSha256 } from './util/crypto/crypto.js'
-export * from './util/encoding.js'
-export * from './util/util.js'
